Create NCUserBaseUrl.swift
